+++ python/GafferUSDUI/__init__.py
@@ -37,5 +37,6 @@
 from . import USDAttributesUI
 from . import USDLayerWriterUI
 from . import USDShaderUI
+from . import USDLightUI
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferUSDUI" )