+++ Changes.md
@@ -5,10 +5,13 @@ Features
 --------
 
 - USDShader : Added a node for loading shaders from USD's `SdrRegistry`. This includes shaders such as `UsdPreviewSurface` and `UsdUVTexture`, which are now available in the `USD/Shader` section of the node menu.
+- USDLight : Added a node for defining UsdLux lights. This is available from the `USD/Light` section of the node menu.
 
 Improvements
 ------------
 
+- LightEditor : Added section displaying UsdLux shadow parameters.
+- Cycles : Added support for UsdLux lights.
 - LightTool : Added support for editing animated plugs.
 
 Fixes